requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
numpy==1.26.4
pillow==10.3.0
tensorflow==2.18.0
ultralytics==8.2.26
opencv-python-headless==4.10.0.82
requests==2.32.3
python-multipart==0.0.9
torch==2.1.2
